--- src/components/myPage/myProfile/Avatar.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Avatar = () => {
-  return <div>Avatar</div>;
-};
-
-export default Avatar;
